ASAGW-56 No Config Client in GWControllerJava
This change is not finished yet. Just implemented the functions that
interface with config service. Still need to implement setting and
storing random passcodes and displaying the lock icon on a gatewayagent
at the correct time. Also there may be extra imports in the java
files, because I was working on another bug too.

Change-Id: Ief90f2f231265be0c8fe95a3f44e0bf7cde65407
Signed-off-by: George Tang <gtang@affinegy.com>