ASAGW-76 JavaGWController Bus cleanup
Bus was not disconnected on closing of the application. Now because it
disconnects in the onDestroy of the DiscoveredGatewaysActivity, it
needed to be connected in the onCreate of the
DiscoveredGatewaysActivity. Since the onCreate of the application is not
called when the application is still in memory, there would be
nullpointer exceptions when reopening the app after pressing back out
of it.

Conflicts:
	java/GatewayController/src/org/alljoyn/gatewaycontroller/activity/BaseActivity.java
	java/GatewayController/src/org/alljoyn/gatewaycontroller/activity/DiscoveredGatewaysActivity.java

Change-Id: Ie3a4363ff3d806b6d43d87ae14f6f5e63a39482c
Signed-off-by: George Tang <gtang@affinegy.com>
Signed-off-by: Josh Spain <jspain@affinegy.com>