ASAGW-56 Implement ConfigClient JavaGWController
The gateway controller now sets random passcodes of announced gateway agents
and stores the random passcode to SharedPreferences with a key value of
the appid. So you can retrieve the passcode from SharedPreferences via
the appid.  Right now, it will always set the passcode of a gwagent
(even if the passcode has been set before) to catch the behavior
of changing the passcode from the configuration file
instead of checking if the passcode has been set before.

Change-Id: I3a1161fe9ef3169522978b4472e6806bc147e858
Signed-off-by: George Tang <gtang@affinegy.com>